install fails earlier when no binaries can be found
this adds a new `CompileMode::Install` variant, which still mostly acts the same as
`CompileMode::Build`, except it filters for only bin targets in `filter_default_targets`.

the result is that root-level resolved libs are not automatically added to the proposal list unless
a valid bin target depends on them, so they won't get accidentally built.

now when you try to install a crate with no satisfied bin targets, cargo will skip the meat of the
compile phase and give the same error message it currently gives when it finds no bin targets at
all.

fixes #8970